requirement.txt
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
defusedxml==0.7.1
Django==5.2
django-debug-toolbar==5.2.0
django-environ==0.12.0
django-filter==25.1
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-nested-routers==0.94.2
drf-yasg==1.21.10
factory_boy==3.3.3
Faker==37.1.0
idna==3.10
importlib_resources==6.5.2
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
mypy_extensions==1.1.0
oauthlib==3.2.2
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.26.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
swagger-spec-validator==3.0.4
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
